backend/pyproject.toml
[project]
name = "FastAPI-Practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
]

[dependency-groups]
dev = [
    "locust>=2.37.0",
    "ruff>=0.11.6",
]
test = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
]


[tool.ruff]
indent-width = 4
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
addopts = "-v --disable-warnings"
